requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.8.3
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==4.2.16
django-auditlog==3.3.0
django-cors-headers==4.9.0
django-dbbackup==5.0.1
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dlib==20.0.0
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fonttools==4.61.1
frozenlist==1.8.0
h5py==3.16.0
idna==3.11
keras==3.13.2
kiwisolver==1.4.9
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
multidict==6.7.0
namex==0.1.0
numpy==2.3.5
opencv-python==4.13.0.92
openpyxl==3.1.5
optree==0.19.0
packaging==25.0
pandas==2.3.3
pgvector==0.4.2
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
rest-framework-simplejwt==0.0.2
rich==14.3.3
seaborn==0.13.2
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.3
tf-keras==2.15.0
twilio==9.4.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.5
wheel==0.46.3
yarl==1.22.0
